island: Consolidate tests for consts.py
