+++ monkey/tests/unit_tests/monkey_island/cc/server_utils/test_consts.py
@@ -1,5 +1,16 @@
+import platform
+
 from monkey_island.cc.server_utils import consts
 
 
 def test_monkey_island_abs_path():
     assert consts.MONKEY_ISLAND_ABS_PATH.endswith("monkey_island")
+
+
+def test_default_server_config_file_path():
+    if platform.system() == "Windows":
+        server_file_path = f"{consts.MONKEY_ISLAND_ABS_PATH}\\cc\\{consts.SERVER_CONFIG_FILENAME}"
+    else:
+        server_file_path = f"{consts.MONKEY_ISLAND_ABS_PATH}/cc/{consts.SERVER_CONFIG_FILENAME}"
+
+    assert consts.DEFAULT_SERVER_CONFIG_PATH == server_file_path
