--- monkey/tests/unit_tests/monkey_island/cc/test_consts.py
@@ -1,12 +1,0 @@
-import platform
-
-import monkey_island.cc.server_utils.consts as consts
-
-
-def test_default_server_config_file_path():
-    if platform.system() == "Windows":
-        server_file_path = f"{consts.MONKEY_ISLAND_ABS_PATH}\\cc\\{consts.SERVER_CONFIG_FILENAME}"
-    else:
-        server_file_path = f"{consts.MONKEY_ISLAND_ABS_PATH}/cc/{consts.SERVER_CONFIG_FILENAME}"
-
-    assert consts.DEFAULT_SERVER_CONFIG_PATH == server_file_path
